fix: currentPage value.
