+++ ubike-app/src/App.vue
@@ -4,11 +4,12 @@
 
     <UbikeTable 
       :uBikeStops="filtedUbikeStops" 
+      :currentPage="currentPage"
     />
 
     <Pagination 
-      :uBikeStops="filtedUbikeStops"
-      :uBikeStopsSorted="sortedUbikeStops"
+      :uBikeStops="filtedUbikeStops" 
+      :currentPage="currentPage"
     />
   </div>
 </template>
@@ -18,13 +19,15 @@ import Search from './components/Search.vue'
 import UbikeTable from './components/UbikeTable.vue'
 import Pagination from './components/Pagination.vue'
 
+import bus from './components/bus'
+
 export default {
   data () {
     return {
+      bus,
       uBikeStops: [],
       searchText: '',
-      totalPageCount: [],
-      sortedUbikeStops: []
+      currentPage: 1
     }
   },
   components: {
@@ -34,8 +37,9 @@ export default {
   },
   methods: {
     // 用來定義在 vue 實體內使用的函數
-    updateUbikeStops(val) {
-      this.sortedUbikeStops = val;
+    changePage(val) {
+      // 想利用這個function做變更currentPage
+      this.currentPage = val;
     }
   },
   watch: {
@@ -43,20 +47,31 @@ export default {
   },
   computed: {
     // 自動為內部資料計算過的屬性
-    filtedUbikeStops () {
+    filtedUbikeStops() {
       // 過濾搜尋
       return this.uBikeStops.length === 0
         ? []
         : this.uBikeStops.filter(d => d.sna.includes(this.searchText));
     }
   },
-  created () {
+  created() {
     fetch('https://tcgbusfs.blob.core.windows.net/blobyoubike/YouBikeTP.gz')
         .then(res => res.json())
         .then(json => {
           const stops = Object.keys(json.retVal).map(key => json.retVal[key]);
           this.uBikeStops = stops;
         });
+    bus.on('change-page', function(val) {
+      // 這邊取得到值了, 但無法將this.changePage放進來 QQ
+      console.log(val);
+    });
+  },
+  beforeUnmount () {
+    // 銷毀 減少記憶體 要養成良好的習慣
+    // 注意使用名稱 重複
+    bus.off('change-page', function(val) {
+      console.log(val);
+    });
   }
 }
 </script>
